test(\`cqn4sql\`): refactor table alias test suite
Split table-alias.test.js (1553 lines, 96 tests) into 9 focused spec
files under table-alias.test/, following the established pattern from
#1315, #1321, and #1326.
